+++ pkg/catalog/operations/helpers.go
